pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dotctl"
version = "x.x.x"
authors = [
  { name = "Pankaj Jackson", email = "pankajackson@live.co.uk" }
]
description = "Command Line Interface to manage (save, export, import) dot files in the form of profiles."
readme = "README.md"
license = { file = "LICENSE" }

requires-python = ">=3.10"
dependencies = [
  "PyYAML"
]

keywords = [
  "dotfiles",
  "config",
  "settings",
  "backup",
  "dotfile-manager"
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Topic :: System :: Archiving :: Backup",
    "Topic :: Utilities"
]

[project.scripts]
dotctl = "dotctl.main:main"

[project.urls]
Homepage = "https://github.com/pankajackson/dotctl"
"Bug Tracker" = "https://github.com/pankajackson/dotctl/issues"
Source = "https://github.com/pankajackson/dotctl"
